🚨 [DEV-2715] Rumqtt: fixing linter issues.
- Fixing linter issues.
