+++ config/default.toml
@@ -38,7 +38,7 @@ max_segment_count = 10
 # Configuration of server and connections that it accepts
 [v4.1]
 name = "v4-1"
-listen = "0.0.0.0:1885"
+listen = "0.0.0.0:1883"
 next_connection_delay_ms = 1
     [v4.1.connections]
     connection_timeout_ms = 60000
