+++ rumqttd/Cargo.toml
@@ -14,7 +14,7 @@ authors.workspace = true
 [dependencies]
 tokio = { version = "^1.41.1", features = ["rt", "time", "net", "io-util", "macros"]}
 serde = { version = "^1.0.215", features = ["derive"] }
-serde_json = "^1.0.132"
+serde_json = "^1.0.133"
 bytes = { version = "^1.8.0", features = ["serde"] }
 flume = { version = "^0.11.1", default-features = false, features = ["async"]}
 slab = "^0.4.9"
@@ -35,7 +35,7 @@ tracing-subscriber = { version="^0.3.18", features=["env-filter"] }
 metrics = "^0.24.0"
 metrics-exporter-prometheus = { version = "^0.16.0", default-features = false, features = ["http-listener"] }
 clap = { version = "^4.5.21", features = ["derive"] }
-axum = "^0.7.7"
+axum = "^0.7.9"
 rand = "^0.8.5"
 uuid = { version = "^1.11.0", features = ["v4", "fast-rng"] }
 subtle = "^2.6.1"
@@ -55,8 +55,8 @@ allow-duplicate-clientid = []
 
 [dev-dependencies]
 pretty_env_logger = "0.5.0"
-config = "0.14"
-pretty_assertions = "1.4.0"
+config = "0.14.1"
+pretty_assertions = "1.4.1"
 
 [[bin]]
 name = "show_settings"
