+++ rumqttd/src/requests/utils_webhook.rs
@@ -1,6 +1,27 @@
+use bytes::Bytes;
+
+use regex::Regex;
+
 use crate::protocol::Packet;
 
 use super::utils_requests::{self, WebhookPayload};
+
+pub fn add_username_to_topic(original_topic: &str, username: Option<String>) -> String {
+    let mut new_topic = original_topic.to_string();
+    if let Some(username) = username {
+        new_topic = new_topic.replace("/v1.6/", &format!("/v1.6/users/{username}/"));
+        new_topic = new_topic.replace("/v2.0/", &format!("/v2.0/users/{username}/"));
+    }
+    new_topic
+}
+
+pub fn remove_username_from_topic(topic: Bytes) -> Bytes {
+    let re = Regex::new(r"/users/[^/]+").unwrap();
+    let topic_string: String = String::from_utf8(topic.to_vec()).unwrap();
+    let topic_string = re.replace(&topic_string, "");
+    let topic_bytes = topic_string.as_bytes().to_vec();
+    Bytes::from(topic_bytes)
+}
 
 pub async fn call_webhook_from_packet(
     packet: Packet,
@@ -11,9 +32,9 @@ pub async fn call_webhook_from_packet(
     match packet.clone() {
         Packet::Subscribe(subscribe, _props) => {
             for f in subscribe.filters {
-                let topic = f.path;
                 let client_id = client_id.to_string();
                 let username = username.to_owned();
+                let topic = add_username_to_topic(&f.path, username.to_owned());
                 if let Some(webhook_url) = webhook_url.to_owned() {
                     let _response = utils_requests::webhook(
                         &webhook_url,
